move alert and historical alert to folder
