+++ server/task/tasks/log-historical-alerts-task.ts
@@ -1,5 +1,5 @@
 import { App } from "@/server/app";
-import { HistoricalAlert } from "@/server/data/historical-alert";
+import { HistoricalAlert } from "@/server/data/alert/historical-alert";
 import { HISTORICAL_ALERTS } from "@/server/database/models/models";
 import { IntervalScheduler } from "@/server/task/lib/interval-scheduler";
 import { Task } from "@/server/task/lib/task";
